+++ apps/cowswap-frontend/jest.config.mjs
@@ -23,5 +23,6 @@ export default {
     '^wagmi$': require.resolve('wagmi'),
     '^@reown/appkit/react$': '<rootDir>/../../testing/reownMock.ts',
     '^@reown/appkit-adapter-wagmi$': '<rootDir>/src/mocks/reownAdapterMock.ts',
+    '^@reown/appkit-adapter-solana/react$': '<rootDir>/src/mocks/reownSolanaAdapterMock.ts',
   },
 }